Added the Tabular Model
This Notebook provides the complete pipeline for evaluating Metadata and the Engineered features as well as the Models necessary to do so. The important features of this model are the custom partial AUC (pAUC) score that is specific to the task and the K-Folds evaluation metrics. The final models considered were a Voting Ensemble of XGBoost and LightGBM and a PyTorch-TabNet architecture. While evaluation of the models were difficult, the averaged probabilities from the K-Folds model seemed to be working better with the Ensemble discussed previously.
Final Metrics on Cross-Validation sets (No SMOTE)-
pAUC of 5-fold mean predictions: 0.1999848977038308 
Recall of 5-fold mean predictions: 1.0